[perf] Create custom JSON serializer for Component ID.
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/722317
Commit-Queue: Joe Gregorio <jcgregorio@google.com>

+++ perf/go/notify/issuetracker.go
@@ -65,7 +65,7 @@ func (t *IssueTrackerTransport) SendNewRegression(ctx context.Context, alert *al
 			Comment: body,
 		},
 		IssueState: &issuetracker.IssueState{
-			ComponentId: alert.IssueTrackerComponent,
+			ComponentId: int64(alert.IssueTrackerComponent),
 			Priority:    "P2",
 			Severity:    "S2",
 			Reporter: &issuetracker.User{
